% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_definitions.R
\docType{data}
\name{races}
\alias{races}
\title{Dataset for races vs tournaments study}
\format{An object of class \code{data.frame} with 299 rows and 42 columns.}
\usage{
data(races)
}
\description{
Sample: demographics, skill rating, individual outcomes, and assignment to treatment groups
}
\examples{
data(races)
summary(races)
}
\keyword{contests,}
\keyword{races,}
\keyword{tournaments}
