% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_definitions.R
\name{render.xtable}
\alias{render.xtable}
\title{Render xtable}
\usage{
render.xtable(x, add = NULL, ...)
}
\description{
Render xtable objects in nice latex format
}
\examples{
xtab <- xtable(table(sample(1:5, size=20, replace=TRUE)))
render.xtable(xtab)
}
\keyword{xtable}
