% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_definitions.R
\docType{data}
\name{scores}
\alias{scores}
\title{Dataset for races vs tournaments study}
\format{An object of class \code{data.frame} with 1759 rows and 5 columns.}
\usage{
data(scores)
}
\description{
Submissions panel data
}
\examples{
data(scores)
summary(scores)
}
\keyword{contests,}
\keyword{races,}
\keyword{tournaments}
